add: Add book upload pool and book parser pool.
The upload pool will handle the uploading of books, while the parser pool
will parse the books and save them to SQLite.

However, SQLite does not support concurrent writes, so it is necessary to
add a goroutine to handle the saving process.

It's worth noting that concurrent writes may occur on the server. For
example, while the server is handling parsing books and saving them to
SQLite, a new request may enter the server, triggering user creation.

To address this issue, I propose a solution involving two executions named
readExec and writeExec. The readExec will read from SQLite and support
concurrent reads, while the writeExec will write to SQLite and not support
concurrency. All SQL operations will be executed by readExec and writeExec.